+++ src/core_foundation/cf_allocator/mod.rs
@@ -1,5 +1,9 @@
 use super::CFType;
 
+mod callbacks;
+
+pub use callbacks::*;
+
 subclass! {
     /// An allocator object.
     ///